fix: be more explicit about segmented button example page
